pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "QualysPy"
version = "0.5.9.0"
authors = [
  { name="Jordan Barnartt", email="jbarnart@uwaterloo.ca" },
]
description = "A Python wrapper for the Qualys API."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License",
    "Natural Language :: English",
    "Operating System :: POSIX :: Linux",
    "Topic :: Security",
    "Typing :: Typed",
]
dependencies = [
  "httpx",
  "python-dateutil",
  "pydantic",
  "SQLAlchemy",
  "psycopg[binary]",
  "pydantic-xml",
  "python-decouple",
  "lxml",
]

[project.optional-dependencies]
dev = [
  "twine",
  "lxml-stubs"
]

[project.urls]
"Homepage" = "https://github.com/JordanBarnartt/qualyspy"
"Bug Tracker" = "https://github.com/JordanBarnartt/qualyspy/issues"

[tool.pyright]
reportUnboundVariable = false
reportUnknownArgumentType = false
reportPrivateUsage = false
reportUnknownVariableType = false
reportUnknownMemberType = false
reportGeneralTypeIssues = false
reportUntypedBaseClass = false
reportInvalidTypeVarUse = false

[[tool.mypy.overrides]]
module = "sqlalchemy"
ignore_errors = true